api/requirements.txt
astroid==1.6.3
attrs==17.4.0
cached-property==1.4.2
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
Django==2.0.4
django-cors-headers==2.2.0
django-extensions==2.0.7
django-rest-registration==0.3.9
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docker==3.2.1
docker-compose==1.21.0
docker-pycreds==0.2.2
dockerpty==0.4.1
docopt==0.6.2
Faker==0.8.8
gunicorn==19.7.1
idna==2.6
isort==4.3.4
jsonschema==2.6.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
mixer==6.0.1
more-itertools==4.1.0
pluggy==0.6.0
psycopg2-binary==2.7.4
py==1.5.3
PyJWT==1.6.1
pylint==1.8.4
pytest==3.5.0
pytest-django==3.2.1
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
six==1.11.0
text-unidecode==1.2
texttable==0.9.1
urllib3==1.22
websocket-client==0.47.0
whitenoise==3.3.1
wrapt==1.10.11
